code regenerated with latest sdk generation code

+++ modules/ibm-analytics-engine-api/src/main/java/com/ibm/cloud/iaesdk/ibm_analytics_engine_api/v2/model/AnalyticsEngineCreateCustomizationResponse.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2021.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -21,7 +21,7 @@ import com.ibm.cloud.sdk.core.service.model.GenericModel;
 public class AnalyticsEngineCreateCustomizationResponse extends GenericModel {
 
   @SerializedName("request_id")
-  protected Double requestId;
+  protected Long requestId;
 
   /**
    * Gets the requestId.
@@ -30,7 +30,7 @@ public class AnalyticsEngineCreateCustomizationResponse extends GenericModel {
    *
    * @return the requestId
    */
-  public Double getRequestId() {
+  public Long getRequestId() {
     return requestId;
   }
 }
